+++ Enlivengine/EnlivengineExamples/EngineExample.cpp
@@ -82,7 +82,7 @@ public:
 		if (mShowMetrics)
 		{
 			ImGui::Begin("Metrics");
-			ImGui::Text("Entities : %d", mRegistry.size());
+			ImGui::Text("Entities : %d", (int)mRegistry.size());
 			ImGui::Text("Mouse X : %d", (int)getApplication().GetWindow().getCursorPosition().x);
 			ImGui::Text("Mouse Y : %d", (int)getApplication().GetWindow().getCursorPosition().y);
 			ImGui::Text("View X : %d", (int)getApplication().GetWindow().getMainView().getCenter().x);
